requirements.txt
torch>=2.2.0
numpy>=1.24.0
PyYAML>=6.0
safetensors>=0.4.2
tqdm>=4.66.0
fastapi>=0.110.0
uvicorn>=0.25.0
pytest>=8.0.0

